backend/requirements.txt
flask
flask-cors
torch
torchaudio
transformers
librosa
soundfile
openai-whisper
numpy